+++ src/main/java/com/atlauncher/gui/tabs/settings/GeneralSettingsTab.java
@@ -73,6 +73,7 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
     private final JCheckBox rememberWindowSizePosition;
     private final JCheckBox useNativeFilePicker;
     private final JCheckBox useRecycleBin;
+    private final JCheckBox enableArmSupport;
 
     public GeneralSettingsTab() {
         // Language
@@ -513,6 +514,26 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
         useRecycleBin = new JCheckBox();
         useRecycleBin.setSelected(App.settings.useRecycleBin);
         add(useRecycleBin, gbc);
+
+        // Enable ARM Support
+
+        gbc.gridx = 0;
+        gbc.gridy++;
+        gbc.insets = UIConstants.LABEL_INSETS;
+        gbc.anchor = GridBagConstraints.BASELINE_TRAILING;
+        JLabelWithHover enableArmSupportLabel = new JLabelWithHover(GetText.tr("Enable ARM Support?"), HELP_ICON,
+                new HTMLBuilder().center().split(100)
+                        .text(GetText
+                                .tr("Support for ARM devices is still experimental. If you experience issues on an ARM based device, please turn this off."))
+                        .build());
+        add(enableArmSupportLabel, gbc);
+
+        gbc.gridx++;
+        gbc.insets = UIConstants.CHECKBOX_FIELD_INSETS;
+        gbc.anchor = GridBagConstraints.BASELINE_LEADING;
+        enableArmSupport = new JCheckBox();
+        enableArmSupport.setSelected(App.settings.enableArmSupport);
+        add(enableArmSupport, gbc);
     }
 
     @SuppressWarnings("unchecked")
@@ -579,6 +600,7 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
 
         App.settings.useNativeFilePicker = useNativeFilePicker.isSelected();
         App.settings.useRecycleBin = useRecycleBin.isSelected();
+        App.settings.enableArmSupport = enableArmSupport.isSelected();
     }
 
     @Override